Create settings reducer to track language and rating setting. Connect reducer to selector on main page. Closes #10, closes #13, closes #15, closes #16, and closes #24
