+++ client/components/settingsData/index.js
@@ -1,0 +1,2 @@
+export { default as languages } from "./languages";
+export { default as ratings } from "./ratings";
