+++ client/components/settingsData/languages.js
@@ -1,0 +1,36 @@
+const languages = [
+	{ key: "English", text: "English", value: "en" },
+	{ key: "Spanish", text: "Spanish", value: "es" },
+	{ key: "Portuguese", text: "Portuguese", value: "pt" },
+	{ key: "Indonesian", text: "Indonesian", value: "id" },
+	{ key: "French", text: "French", value: "fr" },
+	{ key: "Arabic", text: "Arabic", value: "ar" },
+	{ key: "Turkish", text: "Turkish", value: "tr" },
+	{ key: "Thai", text: "Thai", value: "th" },
+	{ key: "Vietnamese", text: "Vietnamese", value: "vi" },
+	{ key: "German", text: "German", value: "de" },
+	{ key: "Italian", text: "Italian", value: "it" },
+	{ key: "Japanese", text: "Japanese", value: "ja" },
+	{ key: "Chinese Simplified", text: "Chinese Simplified", value: "zh-CN" },
+	{ key: "Chinese Traditional", text: "Chinese Traditional", value: "zh-TW" },
+	{ key: "Russian", text: "Russian", value: "ru" },
+	{ key: "Korean", text: "Korean", value: "ko" },
+	{ key: "Polish", text: "Polish", value: "pl" },
+	{ key: "Dutch", text: "Dutch", value: "nl" },
+	{ key: "Romanian", text: "Romanian", value: "ro" },
+	{ key: "Hungarian", text: "Hungarian", value: "hu" },
+	{ key: "Swedish", text: "Swedish", value: "sv" },
+	{ key: "Czech", text: "Czech", value: "cs" },
+	{ key: "Hindi", text: "Hindi", value: "hi" },
+	{ key: "Bengali", text: "Bengali", value: "bn" },
+	{ key: "Danish", text: "Danish", value: "da" },
+	{ key: "Farsi", text: "Farsi", value: "fa" },
+	{ key: "Filipino", text: "Filipino", value: "tl" },
+	{ key: "Finnish", text: "Finnish", value: "fi" },
+	{ key: "Hebrew", text: "Hebrew", value: "iw" },
+	{ key: "Malay", text: "Malay", value: "ms" },
+	{ key: "Norwegian", text: "Norwegian", value: "no" },
+	{ key: "Ukranian", text: "Ukranian", value: "uk" },
+];
+
+export default languages;
