+++ client/components/settingsData/ratings.js
@@ -1,0 +1,9 @@
+const ratings = [
+	{ key: "Y", text: "Y", value: "y" },
+	{ key: "G", text: "G", value: "g" },
+	{ key: "PG", text: "PG", value: "pg" },
+	{ key: "PG-13", text: "PG-13", value: "pg-13" },
+	{ key: "R", text: "R", value: "r" },
+];
+
+export default ratings;
